Fixed bug in plan trip screen

+++ DataModeling/C#/DataDelegates/CarsDelegates/CarsGetCarReservationInfo.cs
@@ -31,7 +31,7 @@ namespace DataModeling
                reader.GetInt32("CarRentalID"),
                reader.GetDateTime("RentalDate"),
                reader.GetString("Model"),
-               reader.GetFloat("Price")
+               reader.GetDouble("Price")
                );
         }
     }
